bump to 0.7.0
see changelog for a list of changes

+++ src/import_roms.rs
@@ -63,17 +63,26 @@ pub async fn import_rom(
     romfile_path: &PathBuf,
     progress_bar: &ProgressBar,
 ) -> SimpleResult<()> {
+    progress_bar.println(&format!(
+        "Processing {:?}",
+        romfile_path.file_name().unwrap()
+    ));
+
+    // abort if the romfile is already in the database
+    if find_romfile_by_path(connection, romfile_path.as_os_str().to_str().unwrap())
+        .await
+        .is_some()
+    {
+        progress_bar.println("Already in database");
+        return Ok(());
+    }
+
     let romfile_extension = romfile_path
         .extension()
         .unwrap()
         .to_str()
         .unwrap()
         .to_lowercase();
-
-    progress_bar.println(&format!(
-        "Processing {:?}",
-        romfile_path.file_name().unwrap()
-    ));
 
     if ARCHIVE_EXTENSIONS.contains(&romfile_extension.as_str()) {
         import_archive(
@@ -161,7 +170,10 @@ async fn import_archive(
 
         let rom = match find_rom(connection, size, &crc, &system, &progress_bar).await {
             Some(rom) => rom,
-            None => return Ok(()),
+            None => {
+                move_to_trash(connection, system_directory, romfile_path, progress_bar).await?;
+                return Ok(());
+            }
         };
 
         let mut new_name = OsString::from(&rom.name);
@@ -182,6 +194,7 @@ async fn import_archive(
 
     // archive contains multiple files
     } else {
+        let delete = true;
         for sevenzip_info in sevenzip_infos {
             let size: u64;
             let crc: String;
@@ -224,8 +237,12 @@ async fn import_archive(
             create_or_update_romfile(connection, &new_path, &rom).await;
         }
 
-        // delete archive
-        remove_file(romfile_path).await?;
+        // delete archive or trash archive
+        if delete {
+            remove_file(romfile_path).await?;
+        } else {
+            move_to_trash(connection, system_directory, romfile_path, progress_bar).await?;
+        }
     }
 
     Ok(())
@@ -252,7 +269,10 @@ async fn import_chd(
     let (size, crc) = get_file_size_and_crc(&cue_path, &header, &progress_bar, 1, 1).await?;
     let cue_rom = match find_rom(connection, size, &crc, &system, &progress_bar).await {
         Some(rom) => rom,
-        None => return Ok(()),
+        None => {
+            move_to_trash(connection, system_directory, &cue_path, progress_bar).await?;
+            return Ok(());
+        }
     };
 
     let roms: Vec<Rom> = find_roms_by_game_id(connection, cue_rom.game_id)
@@ -276,6 +296,7 @@ async fn import_chd(
 
     if roms.iter().enumerate().any(|(i, rom)| crcs[i] != rom.crc) {
         progress_bar.println("CRC mismatch");
+        move_to_trash(connection, system_directory, romfile_path, progress_bar).await?;
         return Ok(());
     }
 
@@ -311,7 +332,10 @@ async fn import_cso(
     remove_file(&iso_path).await?;
     let rom = match find_rom(connection, size, &crc, &system, &progress_bar).await {
         Some(rom) => rom,
-        None => return Ok(()),
+        None => {
+            move_to_trash(connection, system_directory, romfile_path, progress_bar).await?;
+            return Ok(());
+        }
     };
 
     let mut new_cso_path = system_directory.join(&rom.name);
@@ -337,7 +361,10 @@ async fn import_other(
     let (size, crc) = get_file_size_and_crc(romfile_path, &header, &progress_bar, 1, 1).await?;
     let rom = match find_rom(connection, size, &crc, &system, &progress_bar).await {
         Some(rom) => rom,
-        None => return Ok(()),
+        None => {
+            move_to_trash(connection, system_directory, romfile_path, progress_bar).await?;
+            return Ok(());
+        }
     };
 
     let new_path = system_directory.join(&rom.name);
@@ -399,7 +426,21 @@ pub async fn create_or_update_romfile(
         }
         None => create_romfile(connection, romfile_path).await,
     };
-    update_rom_romfile(connection, rom.id, romfile_id).await;
+    update_rom_romfile(connection, rom.id, Some(romfile_id)).await;
+}
+
+async fn move_to_trash(
+    connection: &mut SqliteConnection,
+    system_directory: &PathBuf,
+    romfile_path: &PathBuf,
+    progress_bar: &ProgressBar,
+) -> SimpleResult<()> {
+    let new_path = system_directory
+        .join("Trash")
+        .join(romfile_path.file_name().unwrap());
+    move_file(romfile_path, &new_path, progress_bar).await?;
+    create_romfile(connection, new_path.as_os_str().to_str().unwrap()).await;
+    Ok(())
 }
 
 async fn move_file(
@@ -855,6 +896,87 @@ mod test {
             romfile.path,
             system_path
                 .join("Test Game (USA, Europe).rom")
+                .as_os_str()
+                .to_str()
+                .unwrap(),
+        );
+        assert!(Path::new(&romfile.path).is_file().await);
+        assert_eq!(rom.romfile_id, Some(romfile.id));
+    }
+
+    #[async_std::test]
+    async fn test_import_other_headered() {
+        // given
+        let _guard = MUTEX.get_or_init(|| Mutex::new(0)).lock().await;
+
+        let test_directory = Path::new("test");
+        let progress_bar = ProgressBar::hidden();
+
+        let db_file = NamedTempFile::new().unwrap();
+        let mut connection = establish_connection(db_file.path().to_str().unwrap()).await;
+
+        let dat_path = test_directory.join("Test System 20210402 (Headered).dat");
+        import_dat(&mut connection, &dat_path, false, &progress_bar)
+            .await
+            .unwrap();
+
+        let tmp_directory = TempDir::new_in(&test_directory).unwrap();
+        set_rom_directory(PathBuf::from(&tmp_directory.path()));
+        let tmp_path = set_tmp_directory(PathBuf::from(&tmp_directory.path()));
+        let system_path = &tmp_path.join("Test System");
+        create_directory(&system_path).await.unwrap();
+        let rom_path = tmp_path.join("Test Game (USA, Europe) (Headered).rom");
+        fs::copy(
+            test_directory.join("Test Game (USA, Europe) (Headered).rom"),
+            &rom_path.as_os_str().to_str().unwrap(),
+        )
+        .await
+        .unwrap();
+
+        let system = find_systems(&mut connection).await.remove(0);
+        let header = find_header_by_system_id(&mut connection, system.id).await;
+
+        // when
+        import_other(
+            &mut connection,
+            &system_path,
+            &system,
+            &header,
+            &rom_path,
+            &progress_bar,
+        )
+        .await
+        .unwrap();
+
+        // then
+        let mut roms = find_roms_with_romfile_by_system_id(&mut connection, system.id).await;
+        assert_eq!(roms.len(), 1);
+        let mut romfiles = find_romfiles(&mut connection).await;
+        assert_eq!(romfiles.len(), 1);
+        let mut games = find_games_by_ids(
+            &mut connection,
+            &roms
+                .iter()
+                .map(|rom| rom.game_id)
+                .collect::<Vec<i64>>()
+                .as_slice(),
+        )
+        .await;
+        assert_eq!(games.len(), 1);
+
+        let game = games.remove(0);
+        assert_eq!(game.name, "Test Game (USA, Europe) (Headered)");
+        assert_eq!(game.system_id, system.id);
+
+        let rom = roms.remove(0);
+        assert_eq!(rom.name, "Test Game (USA, Europe) (Headered).rom");
+        assert_eq!(rom.game_id, game.id);
+
+        let romfile = romfiles.remove(0);
+        assert_eq!(
+            romfile.path,
+            system_path
+                .join("Test Game (USA, Europe) (Headered).rom")
                 .as_os_str()
                 .to_str()
                 .unwrap(),
